+++ apps/mobile/src/app/common/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useAppVersionCheck';
 export * from './useQueryString';
 export * from './useSubscriptionIap';
+export * from './useInitializeDeepLink';